chore(core): refactor + cleanup for plugin utils
