+++ packages/nx/src/project-graph/plugins/get-plugins.ts
@@ -1,21 +1,33 @@
-import { hashObject } from '../../hasher/file-hasher';
-import { readNxJson } from '../../config/nx-json';
-import {
-  loadDefaultNxPlugins,
-  LoadedNxPlugin,
-  loadSpecifiedNxPlugins,
-} from './internal-api';
-import { workspaceRoot } from '../../utils/workspace-root';
+import { join } from 'node:path';
 
+import { shouldMergeAngularProjects } from '../../adapter/angular-json';
+import { PluginConfiguration, readNxJson } from '../../config/nx-json';
+import { hashObject } from '../../hasher/file-hasher';
+import { IS_WASM } from '../../native';
+import { workspaceRoot } from '../../utils/workspace-root';
+import { loadNxPluginInIsolation } from './isolation';
+import { loadNxPlugin } from './in-process-loader';
+
+import type { LoadedNxPlugin } from './loaded-nx-plugin';
+import {
+  cleanupPluginTSTranspiler,
+  pluginTranspilerIsRegistered,
+} from './transpiler';
+
+/**
+ * Stuff for specified NX Plugins.
+ */
 let currentPluginsConfigurationHash: string;
 let loadedPlugins: LoadedNxPlugin[];
 let pendingPluginsPromise:
   | Promise<readonly [LoadedNxPlugin[], () => void]>
   | undefined;
-let cleanup: () => void;
+let cleanup: () => void | undefined;
 
-export async function getPlugins(): Promise<LoadedNxPlugin[]> {
-  const pluginsConfiguration = readNxJson().plugins ?? [];
+export async function getPlugins(
+  root = workspaceRoot
+): Promise<LoadedNxPlugin[]> {
+  const pluginsConfiguration = readNxJson(root).plugins ?? [];
   const pluginsConfigurationHash = hashObject(pluginsConfiguration);
 
   // If the plugins configuration has not changed, reuse the current plugins
@@ -32,20 +44,21 @@ export async function getPlugins(): Promise<LoadedNxPlugin[]> {
     cleanup();
   }
 
-  pendingPluginsPromise ??= loadSpecifiedNxPlugins(
-    pluginsConfiguration,
-    workspaceRoot
-  );
+  pendingPluginsPromise ??= loadSpecifiedNxPlugins(pluginsConfiguration, root);
 
   currentPluginsConfigurationHash = pluginsConfigurationHash;
   const [[result, cleanupFn], defaultPlugins] = await Promise.all([
     pendingPluginsPromise,
-    getOnlyDefaultPlugins(),
+    getOnlyDefaultPlugins(root),
   ]);
   cleanup = cleanupFn;
   loadedPlugins = result.concat(defaultPlugins);
   return loadedPlugins;
 }
+
+/**
+ * Stuff for default NX Plugins.
+ */
 
 let loadedDefaultPlugins: LoadedNxPlugin[];
 let cleanupDefaultPlugins: () => void;
@@ -53,7 +66,7 @@ let pendingDefaultPluginPromise:
   | Promise<readonly [LoadedNxPlugin[], () => void]>
   | undefined;
 
-export async function getOnlyDefaultPlugins() {
+export async function getOnlyDefaultPlugins(root = workspaceRoot) {
   // If the plugins configuration has not changed, reuse the current plugins
   if (loadedDefaultPlugins) {
     return loadedPlugins;
@@ -76,6 +89,144 @@ export async function getOnlyDefaultPlugins() {
 export function cleanupPlugins() {
   pendingPluginsPromise = undefined;
   pendingDefaultPluginPromise = undefined;
-  cleanup();
-  cleanupDefaultPlugins();
+  cleanup?.();
+  cleanupDefaultPlugins?.();
+}
+
+/**
+ * Stuff for generic loading
+ */
+
+function isIsolationEnabled() {
+  // Explicitly enabled, regardless of further conditions
+  if (process.env.NX_ISOLATE_PLUGINS === 'true') {
+    return true;
+  }
+  if (
+    // Explicitly disabled
+    process.env.NX_ISOLATE_PLUGINS === 'false' ||
+    // Isolation is disabled on WASM builds currently.
+    IS_WASM
+  ) {
+    return false;
+  }
+  // Default value
+  return true;
+}
+
+const loadingMethod = isIsolationEnabled()
+  ? loadNxPluginInIsolation
+  : loadNxPlugin;
+
+async function loadDefaultNxPlugins(root = workspaceRoot) {
+  performance.mark('loadDefaultNxPlugins:start');
+
+  const plugins = getDefaultPlugins(root);
+
+  const cleanupFunctions: Array<() => void> = [];
+  const ret = [
+    await Promise.all(
+      plugins.map(async (plugin) => {
+        performance.mark(`Load Nx Plugin: ${plugin} - start`);
+
+        const [loadedPluginPromise, cleanup] = await loadingMethod(
+          plugin,
+          root
+        );
+
+        cleanupFunctions.push(cleanup);
+        const res = await loadedPluginPromise;
+        performance.mark(`Load Nx Plugin: ${plugin} - end`);
+        performance.measure(
+          `Load Nx Plugin: ${plugin}`,
+          `Load Nx Plugin: ${plugin} - start`,
+          `Load Nx Plugin: ${plugin} - end`
+        );
+
+        return res;
+      })
+    ),
+    () => {
+      for (const fn of cleanupFunctions) {
+        fn();
+      }
+      if (pluginTranspilerIsRegistered()) {
+        cleanupPluginTSTranspiler();
+      }
+    },
+  ] as const;
+  performance.mark('loadDefaultNxPlugins:end');
+  performance.measure(
+    'loadDefaultNxPlugins',
+    'loadDefaultNxPlugins:start',
+    'loadDefaultNxPlugins:end'
+  );
+  return ret;
+}
+
+async function loadSpecifiedNxPlugins(
+  plugins: PluginConfiguration[],
+  root = workspaceRoot
+): Promise<readonly [LoadedNxPlugin[], () => void]> {
+  performance.mark('loadSpecifiedNxPlugins:start');
+
+  plugins = await normalizePlugins(plugins, root);
+
+  const cleanupFunctions: Array<() => void> = [];
+  const ret = [
+    await Promise.all(
+      plugins.map(async (plugin) => {
+        const pluginPath = typeof plugin === 'string' ? plugin : plugin.plugin;
+        performance.mark(`Load Nx Plugin: ${pluginPath} - start`);
+
+        const [loadedPluginPromise, cleanup] = await loadingMethod(
+          plugin,
+          root
+        );
+
+        cleanupFunctions.push(cleanup);
+        const res = await loadedPluginPromise;
+        performance.mark(`Load Nx Plugin: ${pluginPath} - end`);
+        performance.measure(
+          `Load Nx Plugin: ${pluginPath}`,
+          `Load Nx Plugin: ${pluginPath} - start`,
+          `Load Nx Plugin: ${pluginPath} - end`
+        );
+
+        return res;
+      })
+    ),
+    () => {
+      for (const fn of cleanupFunctions) {
+        fn();
+      }
+      if (pluginTranspilerIsRegistered()) {
+        cleanupPluginTSTranspiler();
+      }
+    },
+  ] as const;
+  performance.mark('loadSpecifiedNxPlugins:end');
+  performance.measure(
+    'loadSpecifiedNxPlugins',
+    'loadSpecifiedNxPlugins:start',
+    'loadSpecifiedNxPlugins:end'
+  );
+  return ret;
+}
+
+async function normalizePlugins(plugins: PluginConfiguration[], root: string) {
+  plugins ??= [];
+
+  return [...plugins];
+}
+
+function getDefaultPlugins(root: string) {
+  return [
+    join(__dirname, '../../plugins/js'),
+    ...(shouldMergeAngularProjects(root, false)
+      ? [join(__dirname, '../../adapter/angular-json')]
+      : []),
+    join(__dirname, '../../plugins/package-json'),
+    join(__dirname, '../../plugins/project-json/build-nodes/project-json'),
+  ];
 }
