+++ packages/nx/src/project-graph/plugins/index.ts
@@ -2,5 +2,6 @@ export * from './public-api';
 
 // export * from './get-plugins';
 
-export { readPluginPackageJson, registerPluginTSTranspiler } from './loader';
+export { readPluginPackageJson } from './in-process-loader';
+export { registerPluginTSTranspiler } from './transpiler';
 export { createNodesFromFiles } from './utils';
