+++ packages/nx/src/project-graph/plugins/isolation/index.ts
@@ -1,6 +1,6 @@
 import { workspaceRoot } from '../../../utils/workspace-root';
-import { PluginConfiguration } from '../../../config/nx-json';
-import { LoadedNxPlugin } from '../internal-api';
+import type { PluginConfiguration } from '../../../config/nx-json';
+import type { LoadedNxPlugin } from '../loaded-nx-plugin';
 import { loadRemoteNxPlugin } from './plugin-pool';
 
 export async function loadNxPluginInIsolation(
